Issue #5 Databases refactoring step2.
