+++ cats-service/build.gradle
@@ -3,24 +3,29 @@ plugins {
     id 'org.springframework.boot' version '3.0.1'
     id 'io.spring.dependency-management' version '1.1.0'
     id "org.springdoc.openapi-gradle-plugin" version "1.6.0"
+    // ToDo: implement profiles
+    id "org.kordamp.gradle.profiles" version "0.49.0"
 }
 
 group 'ua.cats'
-version 'unspecified'
+version '0.1.0-SNAPSHOT'
 
 configurations {
     compileOnly {
         extendsFrom annotationProcessor
     }
 }
+
 repositories {
     mavenCentral()
 }
+
 dependencyManagement {
     imports {
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:2022.0.0"
     }
 }
+
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
@@ -34,6 +39,7 @@ dependencies {
     runtimeOnly 'org.postgresql:postgresql'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 }
+
 tasks.named('test') {
     useJUnitPlatform()
 }
