+++ cats-service/src/main/java/ua/cats/persistance/entities/Cat.java
@@ -1,4 +1,4 @@
-package ua.cats.model;
+package ua.cats.persistance.entities;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.persistence.*;
@@ -15,14 +15,14 @@ import lombok.*;
 public class Cat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     @Column(name = "name", nullable = false, unique = true)
     private String name;
-    private double age;
+    private Double age;
     private String color;
-    private double normalSpeed;
-    private double jumpSpeed;
-    private int agility;
-    private double eyesight;
+    private Double normalSpeed;
+    private Double jumpSpeed;
+    private Integer agility;
+    private Double eyesight;
 
 }
