+++ cats-service/src/main/java/ua/cats/persistance/repositories/CatRepository.java
@@ -1,7 +1,7 @@
-package ua.cats.repository;
+package ua.cats.persistance.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ua.cats.model.Cat;
+import ua.cats.persistance.entities.Cat;
 
 public interface CatRepository extends JpaRepository<Cat,Long> {
 }
