+++ cats-service/src/main/resources/application.yaml
@@ -1,26 +1,29 @@
-server:
-  port: ${PORT:8091}
-
 spring:
+  profiles:
+    active: dev
   application:
     name: cats-service
   jpa:
     show-sql: true
-    #      ToDo:set none in production
     hibernate:
-      ddl-auto: update
-    #          ddl-auto: "validate"
+      ddl-auto: "validate"
     properties:
       hibernate:
         format_sql: true
         dialect: org.hibernate.dialect.PostgreSQLDialect
   datasource:
-    url: "jdbc:postgresql://${DATABASE_HOST_IP:localhost}:${DATABASE_HOST_PORT:5433}/catsdb"
+    url: "jdbc:postgresql://${DATABASE_HOST_IP:postgres-cats}:${DATABASE_HOST_PORT:5432}/cats_database"
     username: "postgres"
     password: "12345678"
     driver-class-name: "org.postgresql.Driver"
-  logging:
-    level:
-      org.springframework: INFO
-      org.hibernate: DEBUG
-      ua.cats: DEBUG
+
+server:
+  port: ${PORT:8091}
+
+logging:
+  level:
+    org.springframework: INFO
+    org.hibernate: DEBUG
+    #    org.hibernate.SQL: DEBUG
+    #    org.hibernate.type: TRACE
+    ua.cats: DEBUG
