+++ mice-service/src/main/resources/application.yaml
@@ -8,18 +8,18 @@ spring:
     show-sql: true
     #      ToDo:set none in production
     hibernate:
-      ddl-auto: update
-    #          ddl-auto: "validate"
+      ddl-auto: "validate"
     properties:
       hibernate:
         format_sql: true
         dialect: org.hibernate.dialect.PostgreSQLDialect
   datasource:
-    url: "jdbc:postgresql://${DATABASE_HOST_IP:localhost}:${DATABASE_HOST_PORT:5432}/micedb"
+    url: "jdbc:postgresql://${DATABASE_HOST_IP:localhost}:${DATABASE_HOST_PORT:5434}/mice_database"
     username: "postgres"
     password: "12345678"
     driver-class-name: "org.postgresql.Driver"
-  logging:
+
+logging:
     level:
       org.springframework: INFO
       org.hibernate: DEBUG
